Remove min_pyver=3.9
