+++ tests/functional/d/deprecated/deprecated_module_py39_earlier_pyversion.rc
@@ -2,5 +2,4 @@
 py-version=3.8
 
 [testoptions]
-min_pyver=3.9
 max_pyver=3.10
