+++ tests/functional/ext/typing/typing_deprecated_alias.rc
@@ -1,8 +1,3 @@
 [main]
 py-version=3.9
 load-plugins=pylint.extensions.typing
-
-[testoptions]
-min_pyver=3.9
-
-[typing]
