--- tests/functional/g/generic_alias/generic_alias_mixed_py39.rc
@@ -1,2 +1,0 @@
-[testoptions]
-min_pyver=3.9
