--- tests/functional/m/method_cache_max_size_none_py39.rc
@@ -1,2 +1,0 @@
-[testoptions]
-min_pyver = 3.9
