--- tests/functional/n/nested_min_max_py39.rc
@@ -1,2 +1,0 @@
-[testoptions]
-min_pyver=3.9
