--- tests/functional/n/non_ascii_name/non_ascii_name_decorator.rc
@@ -1,2 +1,0 @@
-[testoptions]
-min_pyver=3.9
