--- tests/functional/p/postponed/postponed_evaluation_pep585_py39.rc
@@ -1,2 +1,0 @@
-[testoptions]
-min_pyver=3.9
