+++ tests/functional/r/regression_02/regression_5408.rc
@@ -1,3 +1,2 @@
 [testoptions]
-min_pyver=3.9
 except_implementations=PyPy
