+++ tests/functional/u/unnecessary/unnecessary_dunder_call.py
