+++ tests/functional/u/unused/unused_import.txt
@@ -7,8 +7,8 @@ unused-import:11:0:11:51::Unused OrderedDict imported from collections:UNDEFINED
 unused-import:11:0:11:51::Unused deque imported from collections:UNDEFINED
 unused-import:12:0:12:22::Unused import re:UNDEFINED
 unused-import:17:0:17:40::Unused SomeOtherName imported from fake:UNDEFINED
-unused-import:54:0:54:9::Unused import os:UNDEFINED
-unused-import:89:4:89:19::Unused import unittest:UNDEFINED
-unused-import:91:4:91:15::Unused import uuid:UNDEFINED
-unused-import:93:4:93:19::Unused import warnings:UNDEFINED
-unused-import:95:4:95:21::Unused import compileall:UNDEFINED
+unused-import:56:0:56:9::Unused import os:UNDEFINED
+unused-import:91:4:91:19::Unused import unittest:UNDEFINED
+unused-import:93:4:93:15::Unused import uuid:UNDEFINED
+unused-import:95:4:95:19::Unused import warnings:UNDEFINED
+unused-import:97:4:97:21::Unused import compileall:UNDEFINED
