--- tests/functional/u/unused/unused_import_py39.rc
@@ -1,2 +1,0 @@
-[testoptions]
-min_pyver=3.9
