--- tests/functional/u/unused/unused_import_py39.txt
@@ -1,1 +1,0 @@
-unused-import:6:0:6:24::Unused Path imported from pathlib:UNDEFINED
