--- tests/functional/u/unused/unused_name_in_string_literal_type_annotation_py39.rc
@@ -1,2 +1,0 @@
-[testoptions]
-min_pyver=3.9
